+++ src/modules/run/update-triggers/update-triggers.ts
@@ -1,5 +1,5 @@
+import { logger } from "../../../shared";
 import { runTriggers } from "./run-trigger";
-import logger from "../../../shared/logger";
 
 export default async () => {
   logger.info("running put-repository-triggers")