+++ src/modules/vtex/commands/login/login-vtex.ts
@@ -1,26 +1,25 @@
-import { ConfigVtexJson } from './../../../../shared/models/global';
-import log from './../../../../shared/logger';
-import axios from 'axios';
-import chalk from 'chalk';
-import ora from 'ora';
-import { saveVtexConfig } from './util/save-credentials';
+import axios from "axios";
+import chalk from "chalk";
+import ora from "ora";
+import { ConfigVtexJson, logger } from "./../../../../shared";
+import { saveVtexConfig } from "./util/save-credentials";
 
 // Call to get auth information
 const getAuth = async (workspace: string, account: string, token: string) => {
   try {
-    const url = `https://${workspace}--${account}.myvtex.com/exito/token?credentials=${token}?config=${makeid()}`
-    log.debug(url);
+    const url = `https://${workspace}--${account}.myvtex.com/exito/token?credentials=${token}?config=${makeId()}`;
+    logger.debug(url);
     return await axios.get(url);
   } catch (error) {
-    log.debug(error);
-    log.error('Error on get auth token');
+    logger.debug(error);
+    logger.error("Error on get auth token");
     process.exit(1);
   }
 };
 
-function makeid() {
-  var text = '';
-  var possible = 'abcdefghijklmnopqrstuvwxyz';
+function makeId() {
+  var text = "";
+  var possible = "abcdefghijklmnopqrstuvwxyz";
 
   for (var i = 0; i < 5; i++)
     text += possible.charAt(Math.floor(Math.random() * possible.length));
@@ -28,13 +27,13 @@ function makeid() {
   return text;
 }
 
-export default async function (
+export default async function(
   account: string,
   workspace: string,
   email: string,
   token: string
 ) {
-  const spinner = ora('Getting auth token \n').start();
+  const spinner = ora("Getting auth token \n").start();
   spinner.stop();
   const auth = await getAuth(workspace, account, token);
   spinner.start();
@@ -47,7 +46,7 @@ export default async function (
       `auth token to use: ${chalk.redBright(authToken.slice(1, 20))}...`
     );
 
-    log.info(
+    logger.info(
       `Credentials for use ${chalk.redBright(account)} as ${chalk.redBright(
         email
       )} at workspace ${chalk.redBright(workspace)}`
@@ -63,11 +62,11 @@ export default async function (
     };
 
     // 1. Overrite the config file from vtex
-    await saveVtexConfig(options)
+    await saveVtexConfig(options);
 
     spinner.succeed(`Now you logged in Vtex!!`);
   } else {
-    spinner.fail('No token information found.');
+    spinner.fail("No token information found.");
     process.exit(1);
   }
 }