+++ src/modules/vtex/commands/login/util/save-credentials.ts
@@ -1,16 +1,14 @@
+const Configstore = require("configstore");
+import { ConfigVtexJson, logger } from "../../../../../shared";
 
-const Configstore = require('configstore');
-import log from './../../../../../shared/logger';
-import { ConfigVtexJson } from './../../../../../shared/models/global';
-
-export const saveVtexConfig = async (configuration:ConfigVtexJson) => {
+export const saveVtexConfig = async (configuration: ConfigVtexJson) => {
   try {
-    const conf =   new Configstore('vtex');
+    const conf = new Configstore("vtex");
     await conf.clear();
-    conf.all = configuration
-    log.info('vtex json save succefully');
+    conf.all = configuration;
+    logger.info("vtex json save successfully");
   } catch (error) {
-    log.error('Error on save the vtex file');
+    logger.error("Error on save the vtex file");
     process.exit(1);
   }
 };