+++ src/update.ts
@@ -1,5 +1,5 @@
 import updateNotifier from 'update-notifier';
-import * as pkg from '../package.json'
+import * as pkg from '../package.json';
 
 export default function notify() {
   updateNotifier({ pkg, updateCheckInterval: 1000 * 60 * 60 * 1 }).notify({ isGlobal: true })