+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+### v2.8.0 2019-11-19
+
+## Changed
+
+- Create shared folder for export all resources from `models` `utils`.
+- Create new command for execute one command in the projects selected
+
 ### v2.7.8 2019-11-05
 
 ## Added